+++ src/tests/test_resultset.py
@@ -77,7 +77,33 @@ def test_resultset_repr_html(result_set):
         "<th>x</th>\n        </tr>\n    </thead>\n    <tbody>\n        "
         "<tr>\n            <td>0</td>\n        </tr>\n        <tr>\n            "
         "<td>1</td>\n        </tr>\n        <tr>\n            <td>2</td>\n        "
-        "</tr>\n    </tbody>\n</table>"
+        "</tr>\n    </tbody>\n</table>\n"
+        "<span style='font-style:italic;font-size:11px'>"
+        "<code>ResultSet</code> : to convert to pandas, call <a href="
+        "'https://jupysql.ploomber.io/en/latest/integrations/pandas.html'>"
+        "<code>.DataFrame()</code></a> or to polars, call <a href="
+        "'https://jupysql.ploomber.io/en/latest/integrations/polars.html'>"
+        "<code>.PolarsDataFrame()</code></a></span><br>"
+    )
+
+
+@pytest.mark.parametrize(
+    "fname, parameters",
+    [
+        ("head", -1),
+        ("tail", None),
+        ("value_counts", None),
+        ("not_df_function", None),
+    ],
+)
+def test_invalid_operation_error(result_set, fname, parameters):
+    with pytest.raises(AttributeError) as excinfo:
+        getattr(result_set, fname)(parameters)
+
+    assert str(excinfo.value) == (
+        f"'{fname}' is not a valid operation, you can convert this "
+        "into a pandas data frame by calling '.DataFrame()' or a "
+        "polars data frame by calling '.PolarsDataFrame()'"
     )